# openapi.json
{
  "components": {
    "schemas": {
...
      "CloneRequest": {
        "properties": {
          "client_op_id": {
            "anyOf": [
              {
                "maxLength": 36,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Op Id"
          },
          "device_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "slug": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to '{source-slug}-copy[-N]'.",
            "title": "Slug"
          }
        },
        "title": "CloneRequest",
        "type": "object"
      },
...
      "ContainerTypeCreated": {
        "properties": {
          "container_type": {
            "$ref": "#/components/schemas/ContainerTypeRead"
          },
          "replayed": {
            "default": false,
            "description": "True when this is the stored response of an earlier request carrying the same client_op_id; no new movement was recorded.",
            "title": "Replayed",
            "type": "boolean"
          }
        },
        "required": [
          "container_type"
        ],
        "title": "ContainerTypeCreated",
        "type": "object"
      },
      "ContainerTypeEdited": {
        "properties": {
          "cloned": {
            "title": "Cloned",
            "type": "boolean"
          },
          "container_type": {
            "$ref": "#/components/schemas/ContainerTypeRead"
          },
          "replayed": {
            "default": false,
            "description": "True when this is the stored response of an earlier request carrying the same client_op_id; no new movement was recorded.",
            "title": "Replayed",
            "type": "boolean"
          }
        },
        "required": [
          "container_type",
          "cloned"
        ],
        "title": "ContainerTypeEdited",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SlotTemplateRead": {
        "properties": {
          "container_type_id": {
            "title": "Container Type Id",
            "type": "integer"
          },
          "grid_cols": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grid Cols"
          },
          "grid_rows": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grid Rows"
          },
          "materialize_slots": {
            "title": "Materialize Slots",
            "type": "boolean"
          },
          "slot_label_params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Label Params"
          },
          "slot_label_scheme": {
            "title": "Slot Label Scheme",
            "type": "string"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/SlotSpecOut"
            },
            "title": "Slots",
            "type": "array"
          }
        },
        "required": [
          "container_type_id",
          "materialize_slots",
          "grid_rows",
          "grid_cols",
          "slot_label_scheme",
          "slot_label_params",
          "slots"
        ],
        "title": "SlotTemplateRead",
        "type": "object"
      },
...
      "SlotTemplateWritten": {
        "properties": {
          "cloned": {
            "title": "Cloned",
            "type": "boolean"
          },
          "container_type_id": {
            "maximum": 9.223372036854776e+18,
            "minimum": 1.0,
            "title": "Container Type Id",
            "type": "integer"
          },
          "replayed": {
            "default": false,
            "description": "True when this is the stored response of an earlier request carrying the same client_op_id; no new movement was recorded.",
            "title": "Replayed",
            "type": "boolean"
          },
          "template": {
            "$ref": "#/components/schemas/SlotTemplateRead"
          }
        },
        "required": [
          "template",
          "cloned",
          "container_type_id"
        ],
        "title": "SlotTemplateWritten",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Self-hosted electronic-component inventory.",
    "title": "Almagest",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/container-types": {
      "get": {
        "operationId": "list_container_types",
        "parameters": [
          {
            "in": "query",
            "name": "is_seed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Seed"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContainerTypeRead"
                  },
                  "title": "Response List Container Types",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Container Types",
        "tags": [
          "container-types"
        ]
      },
      "post": {
        "operationId": "create_container_type",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContainerTypeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerTypeCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Container Type",
        "tags": [
          "container-types"
        ]
      }
    },
    "/api/container-types/{container_type_id}": {
      "get": {
        "operationId": "read_container_type",
        "parameters": [
          {
            "in": "path",
            "name": "container_type_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Container Type Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerTypeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Container Type",
        "tags": [
          "container-types"
        ]
      },
      "patch": {
        "description": "Edit a type. A seed is read-only, so editing one clones it first \u2014\nsee the module docstring for why this, unlike `PartUpdate`, needs the\nidempotency guard.",
        "operationId": "update_container_type",
        "parameters": [
          {
            "in": "path",
            "name": "container_type_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Container Type Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContainerTypeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerTypeEdited"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Container Type",
        "tags": [
          "container-types"
        ]
      }
    },
    "/api/container-types/{container_type_id}/clone": {
      "post": {
        "description": "\"This cabinet is identical to that one\" \u2014 the explicit form of clone,\nwith no edit attached. Works on any type, seed or not.",
        "operationId": "clone_container_type",
        "parameters": [
          {
            "in": "path",
            "name": "container_type_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Container Type Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerTypeCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clone Container Type",
        "tags": [
          "container-types"
        ]
      }
    },
    "/api/container-types/{container_type_id}/slot-template": {
      "get": {
        "description": "The type's effective layout \u2014 generated or materialised, indistinguishably.",
        "operationId": "read_slot_template",
        "parameters": [
          {
            "in": "path",
            "name": "container_type_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Container Type Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotTemplateRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Slot Template",
        "tags": [
          "container-types"
        ]
      },
      "put": {
        "description": "Save the canvas. A seed clones first (see the module docstring); the\ngrid then materialises **unless** `slots` is exactly what the generator\nwould already produce for the (possibly just-updated) canvas size/scheme,\nin which case nothing is written at all.",
        "operationId": "write_slot_template",
        "parameters": [
          {
            "in": "path",
            "name": "container_type_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Container Type Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlotTemplateWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotTemplateWritten"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write Slot Template",
        "tags": [
          "container-types"
        ]
      }
    },
...
  }
}